VirtualBox backend
Charon can now instantiate VirtualBox instances on the local machine.
This is useful for doing test deployments.  The physical machine spec
should just say:

  { deployment.targetEnv = "virtualbox"; };

This will cause Charon to create a VirtualBox instance corresponding
to the logical machine, if it does not already exist.  The instances
can talk to each other and to the host using VirtualBox's host-only
networking.