Assign an internal index to each machine
Each machine now has an internal index that uniquely defines it within
a deployment.  They're assigned in ascending order.  Adding/removing
machines to/from a network doesn't change the index of
existing/remaining machines.  This is useful for generated numbers
such as VPN IP addresses or tunnel device IDs that ideally shouldn't
change when machines are added or removed.