hetzner: Refactor using hetznerNixOpsInstaller.
This now uses our new NixOps Hetzner bootstrap helper, introduced by
NixOS/nixpkgs@2fd13d8fcafdc9291c385787bb8ef4778d3d814d.

Specific details about how it works can be found in the longDescription of the
package.

From our (NixOps) side, we just have to build the bootstrapper, execute it
(which emits a tar stream) and just pipe the output to tar on the remote side,
which should get us the partitioner available in /usr/bin/nixpart and a script
called "activate-remote".

So, we partition just like before and after we're done with setting up /mnt, we
just run "activate-remote" and we're ready for the deployment.

One issue that's still left is the wrapper for the activation script, but that's
going to change soon.

Also, we are now using gzip instead of xz, because it is less CPU hungry and
most older systems in the wild should have it as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>