hetzner: Configure IPv6 addresses by default.
Unfartunately this only applies for the first (eth0) interface unless we
have found a better and more flexible way of assigning the corresponding
subnets.

The way this is done is by appending "ip -6" commands to
networking.localCommands, which certainly isn't nice either, so it might
be a good idea to have NixOS options for configuring static IPv6
addresses and subnets.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>