hetzner-bootstrap: Don't write temporary TAR file.
This is a killer for VM tests, because we either need to create a
temporary virtio mount/disk image or reserve enough RAM so that the TAR
file can fit into a tmpfs.

However, both solutions are not feasible in our case, so we don't use a
temporary file anymore. The way we avoid it is by streaming the base TAR
archive with a prepended file size, followed by the second archive. We
need to do this, because neither GNU nor BSD TAR have a way to
concatenate archives from stdin and to stdout.

For the stream we're now using the following format:

<size1>:<data1><data2>

Common in protocols is to use <size1>:<data1><size2>:<data2>, but we
don't use the size for <data2>, because at that moment we don't even
know it and calculating the size would result in either wasting memory
or - well - temporary files again.

Unfortunately, extraction now is way more complex than we wanted it in
the first place, but unless someone has a better idea on how to
concatenate TAR archives (or any similar archive format) on the fly,
we're going to stick with this solution.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>