tests/hetzner: Factor out APT repository creation.
We now do this properly by also signing the repository with a valid GPG
key instead of trying to force APT into "insecure" mode.

So far we had repository creatien in the derivation creating the rescue
image, which now is split off because we want to trim down build time of
the rescue image to only the essential parts. That way it's easier to
work on and/or debug the rescue image creation.

As for the repository creation itself, it's now comprised of three
derivations:

 * Snakeoil GPG key
 * Snakeoil GPG key Debian package
 * The finished APT repository

Right now I'm not sure whether we can avoid the keyring package, but for
the matter of completeness it's good to have it (especially if we want
to generalize it for <nixpkgs> someday).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>