backends/hetzner: Fix wrong rstrip()
Thanks a lot to @nh2 for spotting this. Of course we want to strip all
the whitespace from the *left* up to the first comment character instead
of to the *right*.

In practice this shouldn't have any implication because first of all
nixos-generate-config right now doesn't use spaces before comments and
second the worst case that could happen is that hw_info is longer than
it should be (at least if I'm not forgetting about something).

Nevertheless however, let's still fix this because it will confuse
readers of the code and it's also ugly.

Signed-off-by: aszlig <aszlig@nix.build>
Reported-by: @nh2